refactor: StockRealTimeRegimePolicy 파일이동
